RobustRAG_2/requirements.txt
chromadb==0.4.22
pandas>=2.0.0
python-dotenv>=1.0.0
torch>=2.0.0
langchain>=0.1.0
langchain_text_splitters>=0.0.1
transformers>=4.36.0
sentence-transformers>=2.2.2
